[PT Run] Renamed BrowserInfo to DefaultBrowserInfo and made it static
